requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
boto3==1.38.23
botocore==1.38.23
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.2
distro==1.9.0
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
mangum==0.19.0
numpy==2.2.6
openai==1.82.0
packaging==25.0
pandas==2.2.3
platformdirs==4.3.8
pyarrow==18.1.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
s3transfer==0.13.0
six==1.17.0
slack_sdk==3.35.0
sniffio==1.3.1
snowflake-connector-python==3.15.0
sortedcontainers==2.4.0
starlette==0.46.2
tabulate==0.9.0
tomlkit==0.13.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
